JavaFX may be the most infuriating thing I've ever had the displeasure of working with.
Added movement events, so that player one is now completely controllable ( will turn off/make toggleable but it was easy to test methods with this setup)

And fixed up a ton of tiny organizational things. Next step is to start turning all the monster methods into nicer, packaged, more specifically purposed methods.

Then last is just to make a simulate() command within gamelogic so that we can run that as a tester.